requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
email-validator==2.2.0
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
email-validator==2.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
wtforms-sqlalchemy
